backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
PyJWT==2.9.0
python-dotenv==1.0.1
openai>=1.40.0
httpx>=0.27.0
pydantic>=2.0
pymupdf>=1.24.0
rank-bm25>=0.2.2
sentence-transformers>=2.4.0
faiss-cpu>=1.8.1
numpy>=1.24.0